pyproject.toml
[project]
name = "korean-pii"
version = "1.0.0"
description = "korean personally identifiable information text and image masking application"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "fastapi==0.121.3",
    "gunicorn==23.0.0",
    "paddleocr==3.2.0",
    "paddlepaddle==3.2.2",
    "pyyaml==6.0.2",
    "pillow==12.0.0",
    "presidio-analyzer==2.2.360",
    "presidio-anonymizer==2.2.360",
    "python-multipart==0.0.20",
    "spacy==3.8.11",
    "torch==2.7.0; platform_system == 'Linux' and platform_machine == 'x86_64'",
    "transformers==4.57.1",
    "uvicorn==0.38.0",
    "onnxruntime==1.23.2",
    "onnx==1.19.1",
    "phonenumbers==9.0.18",
    "protobuf==6.33.1",
]